+++ src/js/IN_globalFilter.js
@@ -5,4 +5,4 @@ s.onload = function() {
     s.parentNode.removeChild(s);
 };
 
-InjectMulti(['globalFilter.js', 'filters/default.js', 'filters/level.js', 'filters/craftable.js']);
+InjectMulti(['globalFilter.js', 'filters/default.js', 'filters/level.js', 'filters/craftable.js','LaunchFilter.js']);